* 'master' of github.com:tensorflow/tensorflow: (10459 commits)
  Go: Update generated wrapper functions for TensorFlow ops.
  Improve layout related multi-output fusion heuristic.
  Go: Update generated wrapper functions for TensorFlow ops.
  compat: Update forward compatibility horizon to 2020-03-27
  Go: Update generated wrapper functions for TensorFlow ops.
  Fixed a wrong reference to a variable.
  Remove a duplicate check.
  Add the target spec and cpu target spec
  Tweak passes: fuse lstms & inline & dce before lower tensor list ops.
  Adds asymmetric quantized inputs for hybrid ops in future models.
  Add option to enable tfrt eager context
  Simply put some utilities of gpu model builder into a separate header file. Such utilities have been tested in the model builder unit test itself.
  Add a benchmark for the index_lookup adapt call.
  Adds asymmetric quantized inputs for hybrid ops in future models.
  Add AbstractContextInterface to abstract runtime
  Add fused RNN/LSTM ops as legal ops in lower tensor list pass.
  More logical block size selection for convolution.
  Optimize reduce float mean for reduce on the last dimension case.
  arena_planner: Create tensor_order vector directly instead of using tensors_set
  Fall back to ShuffleDatasetV1 when no seed generator is present.
  ...

+++ tensorflow/core/framework/function_handle_cache.cc
@@ -22,7 +22,9 @@ namespace tensorflow {
 namespace data {
 
 FunctionHandleCache::FunctionHandleCache(FunctionLibraryRuntime* lib)
-    : lib_(lib), state_handle_(strings::Printf("%lld", random::New64())) {}
+    : lib_(lib),
+      state_handle_(
+          strings::Printf("%lld", static_cast<long long>(random::New64()))) {}
 
 FunctionHandleCache::~FunctionHandleCache() {
   Status s = Clear();
